+++ dump_manager_bmcstored.hpp
@@ -101,33 +101,6 @@ class Manager : public phosphor::dump::Manager
     EventPtr eventLoop;
 
   protected:
-    /** @brief sd_event_add_child callback
-     *
-     *  @param[in] s - event source
-     *  @param[in] si - signal info
-     *  @param[in] userdata - pointer to Watch object
-     *
-     *  @returns 0 on success, -1 on fail
-     */
-    static int callback(sd_event_source*, const siginfo_t* si, void* entry)
-    {
-        // Set progress as failed if packaging return error
-        if (si->si_status != 0)
-        {
-            log<level::ERR>("Dump packaging failed");
-            if (entry != NULL)
-            {
-                reinterpret_cast<phosphor::dump::Entry*>(entry)->status(
-                    phosphor::dump::OperationStatus::Failed);
-            }
-        }
-        else
-        {
-            log<level::INFO>("Dump packaging completed");
-        }
-        return 0;
-    }
-
     /** @brief Calculate per dump allowed size based on the available
      *        size in the dump location.
      *  @returns dump size in kilobytes.
